switch values with names.
I ran into the same issue and it took me quite a while to figure it out.  Basically in my case I was using the column values instead of the column names in my code.  I've updated the code take a look and let me know if this works for you.  For example instead of using group_value use group and instead of using air_date use show.